backend/foodgram/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
attrs==19.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.0.5
django-debug-toolbar==2.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-extra-fields==3.1.1
filelock==3.0.12
flake8==3.8.4
gunicorn==20.1.0
idna==3.2
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.790
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==20.1
Pillow==8.3.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.1.0
pylint==2.6.0
pyparsing==2.4.6
pytest==6.1.2
pytest-django==3.8.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.2.0
wrapt==1.12.1
zipp==2.2.0